Refactor context imports and implement `CategoriesProductsLayout` for nested routes
