+++ frontend/src/context/CategoriesContext.ts
